backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

sqlalchemy==2.0.30
pymysql==1.1.1
alembic==1.13.1

pydantic==2.7.1
pydantic-settings==2.2.1

python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
